Migrate from RAW to TMA data
With a minimum effort, this commit migrated from the previous logic, which used the raw hw_events values, to the TMA values. In this way, there is no need to implement all the TMA formulas.

Furthermore, there is a global filter to set the labels to be shown instead of relying on the chart.js support that would require selecting the labels on each plot.

This code is a preliminary version.