Improve performance and add plugins for zoom and sync.
The zoom strategy now relies on the chartjs-plugin-crosshair plugin,
allowing sync among charts of the same group. The performance is
increased. Furthermore, the highlight method is not based on a
pre-computed dataset (from tracked samples) and canvas background
highlight but adapt the tracked dataset treating it as another data to be
shown. It is arranged to highlight areas that maximize the tracked line
value (1000) and fill the white area below. The result is the same but
without extra computation.

It is possible to put a side-by-side chart to avoid view dispersion and use
TMAM preset to set metrics filters. Moreover, the color system has been
changed to keep each metric's color independent of the preset or other
selected metrics.